requirements.txt
asgiref==3.6.0
Django==4.1.5
django-crispy-forms==1.14.0
dnspython==2.3.0
pymongo==3.12.3
python-decouple==3.8
pytz==2022.7.1
sqlparse==0.2.4
tzdata==2022.7
